tests/wpt/tests/fenced-frame/fedcm-get-credential.https.html

<!DOCTYPE html>
<title>Test FedCM navigator.credentials.get()</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>

<body>

<script>
promise_test(async () => {
  const key = token();
  attachFencedFrame(
      generateURL('resources/fedcm-get-credential-inner.https.html', [key]));
  // Get the result for the fenced frame.
  const fenced_frame_result = await nextValueFromServer(key);
  assert_equals(
      fenced_frame_result,
      'navigator.credentials.get failed',
      'navigator.credentials.get should fail on fenced frame');
}, 'navigator.credentials.get');
</script>

</body>
